JCLOUDS-1293 Add custom IOException retry handler for AWS-EC2
As all methods use POST we can not use the method to determine if funciton is idempotent.
We therefore set all as idempotent to nullify that check and add a custom
IOException retry handler to determine if commands should be retried on IOException.

The custom hander extends the BackoffLimitedRetryHandler, the current handler, so all other behaviour is not affected.

This does not retry any POST methods unless it's ACTION starts with 'Describe'. These functions are idempotent, and therefore safe to retry.

See JCLOUDS-1293